fix data props
