+++ src/components/Journalists/Journalists.jsx
@@ -2,118 +2,16 @@ import React from "react"
 import { Wrapper } from "./styles"
 import JournalistBox from "../JournalistBox"
 
-const Journalists = () => (
+const Journalists = ({ journalists }) => (
   <Wrapper>
-    <JournalistBox
-      name="Pete Mabazza"
-      date="April 24, 1986"
-      place="Tuguegarao City, Cagayan Province"
-      description="Mabazza, a correspondent of Manila Bulletin, was killed when the Communist Party of the Philippines-New People’s Army ambushed a military convoy in Cagayan Province. He was covering the field operations of the Philippine military when they were caught in the crossfire between the military and guerillas."
-    />
-    <JournalistBox
-      name="Wilfredo Vicoy"
-      date="April 25, 1986"
-      place="Tuguegarao City, Cagayan Province"
-      description="Vicoy suffered internal bleeding caused by shrapnel wounds a day after the Communist Party of the Philippines-New People’s Army ambushed a military convoy he was riding. Like Mabazza, Vicoy was a photojournalist who covered field operations of the Philippine military against communist guerillas."
-    />
-    <JournalistBox
-      name="Virgilio Pacala"
-      date="March 24, 1987"
-      place="San Pablo City, Laguna"
-      description="As a correspondent for the Manila Hotline magazine, Pacala was shot in the back of the head by an unknown gunman while riding a jeepney. Shortly before his death, he rebutted Philippine Constabulary Lt. Col. Victor Corpus’ claim that the Plaza Miranda bombing of 1971 was linked to the New People’s Army (NPA). He had also written an expose on the NPA a few days prior to his murder."
-    />
-
-    <JournalistBox
-      name="Dionisio Perpetuo Joaquin"
-      date="April 12, 1987"
-      place="Olongapo City, Zambales"
-      description="An associate publisher for the Olongapo-Zambales News Forum, Joaquin was at the Whiskey Bar, across from a restaurant where the workers’ strike he had organized was taking place, when he was shot in his sleep. The gunman and his accomplice had been identified as Nicomedes Fabro. Fabro had been hired by Fransisco Dimalanta and Amado Alcala, who were paid off by William Hoge, the owner of the restaurant where the strike Joaquin organized was held. William Hoge and an unidentified accomplice were never arrested and convicted."
-    />
-    <JournalistBox
-      event="dxRA RADYO NI JUAN MASSACRE"
-      name="Narciso Balani, Rogie Zagado, Leo Palo, and Cesar Maglalang"
-      date="August 27, 1987"
-      place="Davao City"
-      description="Maglalang and his colleagues Balani, Zagado, and Palo were killed in the dxRA radio station by suspected communist rebels. Balani was a radio technician; Zagado, a commentator; Palo, an anti-communist crusader who anchored the program Bantay Bayan Kontra Komunista; and Palo, a defeated congressional candidate and commentator."
-    />
-    <JournalistBox
-      name="Martin Castor"
-      date="August 28, 1987"
-      place="Manila"
-      description="Castor was covering a coup attempt as a photojournalist at Pilipino Ngayon when the van he was riding was gunned down by members of a mutinous armed forces faction."
-    />
-    <JournalistBox
-      name="Ramon Noblejas"
-      date="October 4, 1987"
-      place="Tacloban City, Leyte"
-      description="While working for the Manila Broadcasting Company dyVL Aksyon Radyo, Noblejas was gunned down by four assailants. He was known to report on the military’s involvement in corruption, illegal logging and gambling, and human rights abuses."
-    />
-    <JournalistBox
-      name="Leo Enriquez III"
-      date="October 10, 1987"
-      place="Cebu City"
-      description="Enriquez was a reporter affiliated with Kyodo news service, Washington Times, and People’s Journal when he was shot in his home. Police suspected communist rebel groups to be responsible, after the National Democratic Front threatened going after military agents pretending to be journalists. Enriquez had been covering the military since 1983, and his family had survived multiple arson attempts from the New People’s Army."
-    />
-    <JournalistBox
-      name="Noel Miranda"
-      date="March 29, 1988"
-      place="Tagum City, Davao Del Norte"
-      description="As the owner of the Mindanao Scanner, he covered graft and corruption and was known to be actively involved in government efforts to thwart the communist movement. Miranda was shot while walking home."
-    />
-    <JournalistBox
-      name="Ruben R. Manrique"
-      date="August 12, 1988"
-      place="Balanga City, Bataan"
-      description="Manrique was shot three times as he was boarding his car. He was the owner of the Luzon Tribune and renowned for his hard-hitting coverage of local government officials’ misconduct and gambling operators."
-    />
-    <JournalistBox
-      name="Josef Aldeguer Nava"
-      date="October 30, 1988"
-      place="Iloilo City"
-      description="Nava was working for Visayan Life Today and dyRP when he was shot in a cab on the way home. His son cited his work as a journalist to be the motivation for his killing. He was then working on an expose of alleged fictitious loans at the Government Service Insurance System. He also covered drug syndicates."
-    />
-    <JournalistBox
-      name="Severino Arcones"
-      date="October 17, 1989"
-      place="Iloilo City"
-      description="A news anchor at dyFM-Bombo Radyo, Arcones was shot in his home. He hosted a daily segment that tackled matters against the New People’s Army (NPA), having once exposed them to have levied “revolutionary taxes” on businessmen. He had already received threats in the past for his continued attack on the NPA."
-    />
-    <JournalistBox
-      name="Eddie Telan"
-      date="December 1, 1989"
-      place="Quezon City"
-      description="Telan died after having been critically wounded in his coverage of a failed right-wing coup. He was a newscaster."
-    />
-    <JournalistBox
-      name="Enrique Lingan"
-      date="February 4, 1990"
-      place="Lucena City, Quezon"
-      description="Lingan was a columnist affiliated with the Luzon Times and the Midway Star. He died of internal hemorrhage after police officer Felipe Navarro beat him up. A heated exchange between the two ensued after Lingan was taken to the precinct, having been prevented from covering a nude show allegedly supported by the police. The exchange then led to Navarro flooring Lingan and the former banging the latter’s head against the concrete floor."
-    />
-    <JournalistBox
-      name="Joseph “Joe” Kreuger"
-      date="February 6, 1990"
-      place="Pinamalayan, Oriental Mindoro"
-      description="An editor at the Mindoro Weekly Reporter, Kreuger was shot on the way home by an unidentified gunman. Krueger was able to run and the gunman failed at an attempt to shoot him again when barangay tanods and policemen came. He died during operation. He was known as a vocal critic of jueteng operations, having just held a rally a week before his death. In the rally, he named top-ranking military men and public officials who were jueteng protectors."
-    />
-    <JournalistBox
-      name="Reynaldo Catindig Sr."
-      date="May 15, 1990"
-      place="Tumauini, Isabela"
-      description="Catindig worked for the Northern Sierra Madre Express. While several sources list the time and place of his death, there are no other available specifics."
-    />
-    <JournalistBox
-      name="Jean Ladringan"
-      date="July 8, 1990"
-      place="General Santos City"
-      description="Ladringan, publisher and editor at the Southern Star, was killed together with her husband. Ladringan had recently published a piece which incriminated top Muslim officials at the Mindanao State University of graft and corruption."
-    />
-    <JournalistBox
-      name="Nesino Paulin Toling"
-      date="April 14, 1991"
-      place="Ozamiz, Misamis Occidental"
-      description="Publisher and editor for the Panguil Bay Monitor, Toling was shot three times in his office by an unknown assailant. He wrote about the arrest of a suspected drug lord in Northern Mindanao, illegal logging & government corruption. He had been receiving death threats prior to his demise."
-    />
+    {journalists.map((journalist, index) => (
+      <JournalistBox
+        name={journalist.name}
+        date={journalist.date}
+        place={journalist.place}
+        description={journalist.description}
+      />
+    ))}
   </Wrapper>
 )
